Share some gce related code
